fix: structuur
