+++ app/(adoptionprofile)/dog_info.tsx
