+++ app/(adoptionprofile)/index.tsx
@@ -7,7 +7,7 @@ export default function Index() {
 
   useEffect(() => {
     const timer = setTimeout(() => {
-      router.replace("../adoption_profile");
+      router.replace("/adoption_profile");
     }, 3000);
 
     return () => clearTimeout(timer);
