+++ app/(login-register)/index.tsx
@@ -7,7 +7,7 @@ export default function Index() {
 
   useEffect(() => {
     const timer = setTimeout(() => {
-      router.replace("../login_register");
+      router.replace("/login_register");
     }, 3000);
 
     return () => clearTimeout(timer);
