+++ app/(login-register)/login/password/forget_password.tsx
@@ -1,16 +1,23 @@
 import React, { useState } from "react";
-import { View, Text, StyleSheet, TouchableOpacity, TextInput, Alert } from "react-native";
-import { useRouter } from "expo-router";  
+import {
+  View,
+  Text,
+  StyleSheet,
+  TouchableOpacity,
+  TextInput,
+  Alert,
+} from "react-native";
+import { useRouter } from "expo-router";
 import { supabase } from "../../../../lib/supabase";
 
 const ForgetPasswordScreen = () => {
   const router = useRouter();
   const [emailFocus, setEmailFocus] = useState(false);
-  const [email, setEmail] = useState('');
+  const [email, setEmail] = useState("");
   const [loading, setLoading] = useState(false);
 
   // Functie om te controleren of er tekst is ingevoerd
-  const isEmailFilled = email.trim() !== '';
+  const isEmailFilled = email.trim() !== "";
 
   const handleResetPassword = async () => {
     if (!email) {
@@ -21,13 +28,14 @@ const ForgetPasswordScreen = () => {
     setLoading(true);
 
     const { error } = await supabase.auth.resetPasswordForEmail(email, {
-      redirectTo: 'com.anonymous.felicksapp://login-register/login/password/newpassword'
+      redirectTo:
+        "com.anonymous.felicksapp://login-register/login/password/newpassword",
     });
 
     if (error) {
       Alert.alert("Fout", "Deze e-mail is niet geregistreerd.");
     } else {
-      router.push("/login-register/login/password/notification_email");
+      router.push("/login/password/notification_email");
     }
 
     setLoading(false);
@@ -41,22 +49,23 @@ const ForgetPasswordScreen = () => {
       <Text style={styles.label}>E-mail</Text>
       <TextInput
         style={[
-          styles.input, 
-          emailFocus || isEmailFilled ? styles.focusedInput : styles.unfocusedInput
+          styles.input,
+          emailFocus || isEmailFilled
+            ? styles.focusedInput
+            : styles.unfocusedInput,
         ]}
-        placeholder="E-mail" 
+        placeholder="E-mail"
         placeholderTextColor="rgba(151, 184, 165, 0.5)"
         keyboardType="email-address"
-        onFocus={() => setEmailFocus(true)} 
-        onBlur={() => setEmailFocus(false)} 
+        onFocus={() => setEmailFocus(true)}
+        onBlur={() => setEmailFocus(false)}
         onChangeText={setEmail}
         value={email}
       />
-  
+
       <TouchableOpacity style={styles.button} onPress={handleResetPassword}>
         <Text style={styles.buttonText}>VERZEND</Text>
       </TouchableOpacity>
-
     </View>
   );
 };
@@ -65,14 +74,14 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     paddingTop: 100,
-    alignItems: 'center',
+    alignItems: "center",
     padding: 20,
-    backgroundColor: '#FFFDF9',
+    backgroundColor: "#FFFDF9",
   },
   title: {
     fontSize: 23,
     fontWeight: "bold",
-    color: '#183A36',
+    color: "#183A36",
     marginBottom: 60,
   },
   text: {
@@ -80,21 +89,21 @@ const styles = StyleSheet.create({
     fontWeight: "bold",
   },
   button: {
-    backgroundColor: '#97B8A5',
+    backgroundColor: "#97B8A5",
     paddingVertical: 15,
-    borderRadius: 20, 
+    borderRadius: 20,
     marginBottom: 20,
-    width: '97%',
-    alignItems: 'center',
+    width: "97%",
+    alignItems: "center",
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.1,
     shadowRadius: 4,
     marginTop: 20,
   },
   buttonText: {
-    color: '#183A36',
+    color: "#183A36",
     fontSize: 16,
-    fontWeight: 'bold',
+    fontWeight: "bold",
   },
   label: {
     alignSelf: "flex-start",
@@ -107,19 +116,18 @@ const styles = StyleSheet.create({
     width: "100%",
     height: 45,
     borderBottomWidth: 1,
-    borderBottomColor: "#97B8A5", 
+    borderBottomColor: "#97B8A5",
     marginBottom: 25,
     fontSize: 16,
     color: "#183A36",
     paddingLeft: 15,
   },
   focusedInput: {
-    borderBottomColor: '#183A36', 
+    borderBottomColor: "#183A36",
   },
   unfocusedInput: {
-    borderBottomColor: "#97B8A5", 
+    borderBottomColor: "#97B8A5",
   },
-  
 });
 
 export default ForgetPasswordScreen;
