+++ app/index.tsx
@@ -8,7 +8,8 @@ export default function Index() {
   useEffect(() => {
     const timer = setTimeout(() => {
       // router.replace("../login-register/startpage/startpage");
-      router.replace("../login-register/login_register");
+      // router.replace("../login-register/login_register");
+      router.replace("/login_register");
     }, 3000);
 
     return () => clearTimeout(timer);
